refactor: extract transcript extraction logic into TranscriptListExtractor for better modularity
